+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+#### 1.0.0 (07.10.2015)
+
+- Jade was removed
+- grunt-modernizr was removed
+- autoprefixer was added
+- Foundation was updated to newest v5.3.3
+
 #### 0.10.0 (04.10.2015)
 
 - Modernizr fix - now there is a custom build based on [grunt-modernizr](https://github.com/modernizr/grunt-modernizr) task